django-postgres-reactjs/sampleApi/requirements.txt
asgiref==3.3.4
astroid==2.5.6
backoff==1.10.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cors==0.9.11
decorator==5.0.7
defusedxml==0.7.1
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
flake8==3.9.2
Flask==1.1.2
future==0.18.2
gevent==21.1.2
ghp-import==2.0.1
gitdb==4.0.7
GitPython==3.1.17
greenlet==1.0.0
idna==2.10
importlib-metadata==4.5.0
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
jira==3.0.1
keyring==23.0.1
lazy-object-proxy==1.6.0
linecache2==1.0.0
lxml==4.6.3
Markdown==3.3.4
markdown-include==0.6.0
markdown2==2.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
mergedeep==1.3.4
mkdocs==1.2
mkdocs-material==7.1.7
mkdocs-material-extensions==1.0.1
mkdocs-mermaid-plugin==0.1.1
mkdocs-pandoc==0.2.6
msgpack==1.0.2
oauthlib==3.1.1
packaging==20.9
packer.py==0.3.0
pandoc==1.0.2
ply==3.11
py==1.10.0
pycodestyle==2.7.0
pydocstyle==6.1.1
pyfiglet==0.8.post1
pyflakes==2.3.1
Pygments==2.9.0
pylint==2.8.3
pymdown-extensions==8.2
pymetasploit3==1.0.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-nmap==0.6.4
python-vagrant==0.5.15
pythonping==1.0.16
pytz==2021.1
PyYAML==5.4.1
pyyaml-env-tag==0.1
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retry==0.9.2
six==1.15.0
smmap==4.0.0
snowballstemmer==2.1.0
soupsieve==2.2.1
sqlparse==0.4.1
tabulate==0.8.9
termcolor==1.1.0
tinycss2==1.1.0
tldextract==3.1.0
toml==0.10.2
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.26.4
virtualbox==2.1.1
watchdog==2.1.2
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.3.0
